restructure tests. Add happy path test
use mockk context for assertions
